chore: separate admin/public routes and small fix

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -6,9 +6,6 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('users', function (Blueprint $table) {
@@ -21,19 +18,18 @@ return new class extends Migration
             $table->unsignedBigInteger('mdj_id')->nullable();
             $table->rememberToken();
             $table->timestamps();
+        });
 
-            // Ajout de la contrainte de clé étrangère
-            $table->foreign('mdj_id')->references('id')->on('mdjs')->onDelete('set null');
+        Schema::table('users', function (Blueprint $table) {
+            if (Schema::hasTable('mdjs')) {
+                $table->foreign('mdj_id')->references('id')->on('mdjs')->onDelete('set null');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            // Suppression de la contrainte de clé étrangère avant de supprimer la table
             $table->dropForeign(['mdj_id']);
         });
